extract getWorkload method to make it easy to test
